VkInfoFriendService/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
gunicorn==20.1.0
drf_yasg==1.21.5
